actiongutter: New widget to replace GutterRendererChunkAction
The main reason we're looking to replace the existing
GutterRendererChunkAction is that using gutters for actions (and
specifically using the right-hand gutter) doesn't really work visually,
and conflicts somewhat with using overlay scrolling.

The visual problem is that the gutter is included in rendering + sizing
of the GtkSourceView, but when you look at the widget layout on screen,
you kind of expect other alignment points (e.g., per-pane toolbars,
per-pane infobars) not to cover the gutter space... but they do.
Similarly, the visual hint for scrolling area (light dashed lines) on
the textview appears on the far side of the gutter, but within its
allocation... which again feels wrong.

The conflict with overlay scrolling is much more straightforward.
Overlay scrolling draws over the edge sub-window of the textview, which
is by design the action gutter. This means that scrolling covers the
action gutter (which looks wrong), and also that trying to activate
buttons in the action gutter triggers the overlay scrollbar expansion.