Update source style schemes to have parent schemes
This means that if you enable syntax highlighting on the base schemes,
you'll now get actual syntax highlighting. In other words, these schemes
are no longer just placeholders that we get our themed colours from.

The reason that the dark theme isn't actually the default yet is that
unlike the light theme, the background is kind of crucial, and the text
and background colours aren't actually inherited in our required version
of GtkSourceView; this has already been fixed as GNOME/gtksourceview#37,
but we won't be able to require that for at least one release.