filediff: Rename the local-to-filediff-the-widget action group
This was the cause of a weird problem where no `view`-group actions were
actually activate-able from the context of FileDiff, because the `view`
group *on the widget* was found before the `view` group in the window.

We need the group on the window because that's the only way to get
things to show up in the menu. We want (but don't need) the group on
the FileDiff widget because it's saner scoping, and it Just Works for
our status bars.

An example symptom of this was that while you could use the toolbar
buttons to go to the next change, the shortcut didn't work.