dirdiff: Move custom filter actions to dynamically populated GActions
This really feels like bending `GAction` to do a thing that it wasn't
really intended to do, but I can't think of any other way to make the
appropriate UI happen. It's unpleasant here that we have dynamic action
creation as a window-level concern, but the dynamic action creation
occurs in `DirDiff` itself. I mean... this is correct, but it doesn't
feel great.