vcview: Move all VC-specific actions to GActions
This is the last of the vcview GtkUIManager actions. The UI here is
clearly still extremely placeholder, but in service of getting the
actions into GActions it's good enough.

The popup menu still needs to be reinstated, but that's a task for later
when the remainder of the menu conversion has been done.