Remove custom cut/copy/paste actions
Correctly managing the behaviour of these actions isn't that hard - we
already did a pretty good job in the MeldWindow callbacks - but managing
their sensitivity is an actual nightmare. In order to do this correctly,
you need to hook in to focus changes, then connect to selection-changed
and editability-changed signals (which GtkEditable doesn't even have)
and then manage those signal connections so that you can clean them up
when your focus changes.

I started doing all of the above, and then realised a few things...

Firstly, UIs that I'm using as a model don't really expose these actions
as top-level user-visible items anyway. They're expected to be used via
shortcuts or context menus.

Secondly, the context menu handlers for the widgets where this is
relevant already handle all of the sensitivity logic, and can easily do
so because they only need to run the sensitivity checks when the context
menu is invoked.

So all I'm doing here is removing the custom actions, and we'll see how
we go with just relying on the built-in behaviours of GtkEditable and
GtkSourceView. Maybe we'll need to reinstate some of this, but maybe
we'll get away without.