vcview: Switch up VC selection UI to always show all options
The idea here is that in 99% of cases, users would only have one
repository in a folder, but that meant that we'd show them an annoyingly
inactive combobox with little feedback about what else we might be able
to do, or indeed why there was only one option there.

Now, we show all of our VC options in the combo with the reasons that
they're not available indicated. We still try to maintain the same
default selection behaviour that we had previously.