filediff: Bind our statusbar cursor position with an adapter
This is being moved partly for consistency with our other statusbar-
displayed properties, but more importantly because we weren't always
setting the cursor position on the statusbar.

The `on_cursor_position_changed()` callback tries not to do a bunch of
recalculation when the cursor doesn't change, but this was biting us on
initial load. If the cursor didn't get moved due to a change being found
(for example, if you're diffing blank files...) then we didn't
initialise the line/column display _at all_ until the cursor changed
again.