ui.statusbar: Set a reasonable cursor label width
The initial cursor label allocation is almost always enough to hold a
couple of digits for line and a single digit for column... that's just
where we expect initial cursors to end up. Having the allocation change
when the cursor moves makes the statusbar feel dodgy, and it's frankly
much nicer if all of the widgets stay in a constant place.

The calculation here is a trade-off between line count + length that we
expect users might hit vs. not wanting too much blank space between
the cursor button and other statusbar widgets. The calculation is
essentially the formatted version of having <10000 lines at <1000
characters long... and the -2 is from practical testing, since the
allocated space _easily_ fits more than the desired line counts with
normal system fonts.