chunkmap: New widget to replace the old DiffMap implementation
This is very similar to the DiffMap, but has been rewritten in a more
modern Py/GObject style to do more state-based rendering. As a result,
the manual draw queueing from FileDiff is no longer necessary, and the
widget integrates much more loosely with its parent.

In addition, the mouse behaviour has also been rewritten to handle drags
and uses GtkSourceView's scrolling for animated transitions.