chunkmap: Always queue a draw when new chunks are set
This is unfortunate, but otherwise it's difficult to figure out how to
handle invalidation where our `GtkAdjustment`s don't change.

Without this patch, the easy scenario to see this break is to open a
file comparison and add a character to a line that's currently not
marked as different, and note that the `ChunkMap` doesn't update to
show this line as different.