chunkmap, dirdiff: Use new overview map widget for folder comparisons
This includes the same grid packing as file comparisons and the same
visualisation, as well as the visibility control.

In an ideal world, `DirDiff` would use the existing `chunks` property
on `ChunkMap`. While this is completely do-able, expanding and
collapsing the tree completely ruins performance because of recursive
expand/collapse behaviour, and because in a naive implementation (like
the one we currently use in `TreeViewChunkMap`) the tree model state
must be refetched at each expand/collapse. It's possible to avoid this,
but the caching in `DirDiff` became much too complex for the relative
pay-off of not having the chunk map do these calculations lazily.