chunkmap: Scale textview chunk maps using paired textviews
With this patch we now scale the total height of chunk maps for
textviews according to the maximum adjustment height of the views next
to them.

The idea behind this change is that it's misleading to show maps that
are next to each other as being the same size. Comparing files of
different sizes should indicate that this is the case, and insertions
will offset the diffs.