chunkmap: Draw a background with theme colours and update handle
On most documents with a normal number of changes, it was very easy to
for a couple of small changes to be hard to see on the map, because we
didn't draw a background for the widget. This was initially intentional
because it makes it feel a bit less cluttered, but in this case the
lack of clutter makes actual changes stand out less.

This changes the drawing so that we add a map background and outline
based on the GTK+ theme colours. Because it's closely related code, this
also bases the handle colours on the theme, so we now only have the
alpha values as properties (and probably don't even need that).

This theme code *won't* do the right thing if the GtkSourceView scheme
differs massively from the GTK+ theme, but that's a moderately
questionable thing to support anyway. It would also tie the ChunkMap
code more closely to sourceview, which I'd prefer not to do if possible.