bin/meld: Fix missing translations for GTK-loaded resources (#321)
Moving to having template widgets broke translation for those widgets,
because we moved from loading via GtkBuilder with explicit domains from
Python to loading with the domains set in the C library. We'd done
*almost* all the work to bind those, but for some reason hadn't
correctly set `textdomain`.

The Windows libintl change here is purely theoretical; translations
don't work on Windows in the current build, so we can't even test
whether this does the right thing.