Update build helpers for Python 3.8 compatibility (#322)
The `linux_distribution` helper for the platform module has been removed
in Python 3.8, so we need an additional helper to check for the Debian-
style packaging layout.

Really this should be an `install_requires`, but moving our build
helpers to `setuptools` is a not-insignificant risk that I'd rather not
take when we're looking at moving to Meson.