dirdiff: Fix expand/collapse sensitivity for fake empty rows (#342)
We add path-less rows to the model to represent empty folders, which the
existing selection code wasn't handling. With this change, we don't try
to figure out whether the current selection is a folder... we just look
to see whether the model has children, which is both easier and more
correct.