dirdiff: Handle pre-epoch timestamps better (#265)
There's two parts to this handling. The first is that our cell renderer
no longer fails when asked to render timestamps before -2**32-1... it
now just renders them as missing values.

The other part is that now that we kind of support negative timestamps,
I felt like we couldn't keep using -1 as the sentinel. Ideally we'd use
NaN or something, but that doesn't work in GObject currently, so this
current change now uses -2**32-1 as the sentinel. This isn't great,
since it's still a valid timestamp... but it's the best I could come up
with given the limitations.

If we didn't mind giving up on using our `DiffTreeStore.unsafe_set()`
helper, then we could store -2**32-2 or something instead, but
discarding the significant speed-ups from that helper just so that we
can render timestamps that vanishingly-few users will ever see feels
like a bad trade-off.