settings: Add module global accessor to avoid import-time binding
This hasn't been an issue until now because we've been careful to make
sure that settings setup happens early and doesn't import anything else
that might get an uninitialised `meldsettings` object. However, I'm
about to add some imports to this for style scheme initialisation
reasons, so we need this here.

...also it's the correct way to do what we're doing.