meldbuffer, sourceview: Move buffer style scheme handling to sourceview
While this kind of feels like moving it to the wrong place, it's
actually sensible because it's the only sane way to control when we
access the MeldSettings object. With `MeldSourceView` we do this by
delaying until realize(), but `MeldBuffer` is a simple GObject and so
doesn't have the same life-cycle.