style: Rework dark scheme detection based on GtkSourceView style scheme
This makes more sense and works in more scenarios. If a user selects a
dark sourceview syntax highlighting scheme, then that's the background
that's going to get used, regardless of what the GTK+ theme says. The
only scenario in which the GTK+ theme background matters is when the
sourceview scheme doesn't have a background set.