filediff: Fix push-change actions for missing three way chunks (#359)
Way back in 80bfe200 when we added support for pushing change deletions,
removing the "is there anything to push" check had the side effect of
incorrectly allowing the user to try to push a change between panes
where no such change exists.

This patch adds a three-way-comparison-specific check for whether we
actually have a change chunk in the direction that we're trying to push.