filediff: Move auto-merge functionality to main FileDiff class
This isn't ideal logic-wise, but it's the only sane workaround I can
think of to deal with the GtkTemplate support not supporting subclasses
of templated widgets (https://github.com/yousseb/meld/issues/97).

The actual logic separating the two cases isn't much, so even though I
don't like adding the concept of a "mode" to `FileDiff`, this should be
basically fine.