filediff: Handle three-way edge case in viewport-relative scrolling
The scenario here is slightly odd, but in some cases in three-way
comparisons, using the viewport-relative scroll gets stuck going
between two chunks in the middle. What actually happens is that the
scroll down doesn't bring the target to above 50% of the screen
vertically, and so when looking for the "next" chunk from the middle of
the viewport, we either get the current chunk, or in some weirder cases
the previous chunk.

The fix in this patch is to only do viewport-relative scrolling if the
target non-viewport-relative chunk isn't already in the viewport,
falling back to just using the existing logic. This seems to work well
in practice, although it's possible that a future enhancement would be
to do the is-the-chunk-in-the-area check with a restricted area (say
the middle 50% of the viewport) rather than the whole viewport.