misc: Skip missing filter regexes when matching (#442)
When compiling our filter regular expressions, we set the compiled regex
to `None` on compilation errors. While this isn't ideal, it's basically
sensible. However, we then explode when actually trying to match this
filter because we don't ignore the missing compiled regex. This patch
changes that so that we skip these errors.