ui.findbar: Don't hide the find bar when it loses focus (#477)
This was an old UI decision that now feels somewhat dated. Other text
editors have by-and-large switched to explicit dismissal of the find
bar using Escape or similar, which we support. Removing this handling
just removes the behaviour of hiding the find bar when the user clicks
away from it, which feels nicer to use.