ui.statusbar: Ellipsize status bar button labels (#561)
While not generally an issue in two-pane mode, in three-pane comparisons
we can get in to a situation where the status bar button widths severely
restrict the minimum window width. This was technically a problem in
3.20.x, but because there was one fewer button and we didn't provide a
generous minimum allocation for the line/column display, it wasn't as
big of a deal.