sourceview: Remove animation source on MeldSourceView unrealize (#565)
Because we only removed the animation source in do_draw_layer, when the
animation was finished, we could get in to a state where the source view
was unrealized (and therefore no longer getting draw events) while there
was still an animation source running. This could lead to the main
Meld process polling in a tight loop for a source that did nothing other
than constantly requesting draws for a widget that wasn't displayed.

The easy fix here is to remove the source when the widget unrealizes.
Ideally, we'd use a better animation model instead, but... that's a
bigger change.