ui.msgarea: Work around wrapping label allocation issues (#588)
It's unclear what the bug here is, and I haven't been able to narrow it
down well enough for a GTK bug report. In Meld, this problem manifests
as the window getting a large fixed minimum allocation - apparently on
the GtkScrolledWindow child, but I haven't tracked it down - when the
GtkInfoBar has *two* child labels *both* of which use WORD_CHAR or CHAR
wrapping.

Since this bug does not show up if either label has WORD wrapping, here
we're working around the problem by removing the WORD_CHAR wrapping from
the secondary label. Since this wrapping was added to fix a bug in the
primary label display on Windows, this seems like our best option.