meldbuffer: Add a line cache to our Bufferlines helper class (#564)
In the 3.21.x series, we changed this class so that instead of fetching
a text range and splitting it in Python, we instead manually fetched
individual lines using GTK logic. We did this because we have a long
history of GTK and Python disagreeing about what constitutes a line
break. Since the base GtkTextBuffer will always use its own definition,
we have no realistic choice but to go with what it wants.

The problem here is that it turns out that between GTK not really
exposing fast line-oriented APIs, and the number of Python -> GTK
calls we were having to make to get the line data out, this turns out
to be *very* slow.

The solution we've gone with here is to add a Python-side cache on top
of our line accessing logic. This is not entirely dissimilar to our
existing logic that invalidates diff chunks when the underlying buffer
changes.