chunkmap: Improve responsiveness when dragging in ChunkMap
The default behaviour here is to use the GTK scrolling helpers when
e.g., clicking in the map, so that we get nice animated scrolling to a
point. However, this feels very bad when dragging in the ChunkMap,
because the animated delays make the dragging feel very sluggish. Here,
we're doing the simple thing when we're in a drag state and just setting
the adjustment value directly.