recent: Handle bad comparison metadata cases
This happens most often when there is missing data in the comparison,
such as when we have empty comparisons with no location. While we can
try to make the individual comparison types handle this better, it's
also sensible to make the recent handling just ignore this problem,
since it can't possibly record a sensible recent document.