Remove tab labels from app title
Now that we have more complete labelling on individual comparison panes,
whenever we are displaying tabs we end up with indicators showing that
we're comparing e.g., [foo] thing.txt : [bar] thing.txt *three times* on
a single page: the title, the tab label, and the pane labels.

This commit removes the title repetition, so we're now showing it at
most twice.