Rework open/path label UI arrangement
One of the weird things about the path label is that it means that the
UI to open a new file is on the other side of the pane to the UI that
shows you what file you're looking at. This is especially weird for
version control view, since it has additional UI for selecting which
version control system you're using.

This commit switches these up to use a linked button UI for opening and
labelling the files/folders being compared. This recovers a small amount
of action bar space, but more importantly means that the file-y buttons
are in the middle section, leaving the left section for any action-y
buttons that we want. Also, it makes all of the file/folder/version
control layouts consistent.