vcview: Reposition VC selection combo to the right of the action bar
This selector is likely used by a very small subset of users, and with
its repositioning to the middle of the action bar with the file label +
selector, it had way too much prominence. It also had the visual effect
of creating even more clutter on narrower windows because of the density
of widget decoration + text in a small area.