filediff: Check output file override when calculating parent path
If we don't use the output file override, we'll try to change the parent
path from `/` (because the file that's loaded will be in `/tmp`) and
we'll hit a very reasonable error case in the path label display. This
change means that we'll instead use the path that we're going to save
our file to when doing parent calculations.