filediff: Set buffers as modified when using custom save path (#581)
This change is about improving conflict resolution, which is envisaged
as the primary use case for the `--output` flag. The point here is that
if when you use the "save this buffer to a different place than where
you loaded it from" functionality, it's reasonable to expect that the
buffer is going to need to be saved in order for the buffer contents to
end up on disk.

This will change some workflows slightly, in that if you open up a
conflict from Meld's version control view but make no changes, you'll
now be prompted to save the file... but this seems like an acceptable
trade-off.