dirdiff: Make new/deleted file comparison open in two-pane mode (#492)
The issue being addressed here is that once the comparison has been
opened, there's no clear indication of whether it was the left or the
right pane that was missing from the folder comparison.

The down-side to making us always open in two-pane mode is that we lose
the screen real estate of the single-pane view, but the benefits in
consistency feel like they outweigh that consideration. If a user really
wants a full-size view of an added/removed file without the screen
split, the Open Externally action is already available.