dirdiff: Refactor comparison marker to use row references
The goal here is to have the marker survive tree modifications, and also
to make the comparison that we do for rendering emblems somewhat more
correct. Because handling TreeRowReferences is slightly more annoying
than dealing with pure TreeIters, I've moved the marker to be a
NamedTuple to make the rest of our logic appear slightly simpler.