filediff: Apply custom labels to new pathlabel widgets (#662)
This is a behaviour change from where we used to apply labels, in so far
as we used to have labels in the title bar and the tab label, and now
have them in the (sometimes hidden) tab label and the per-pane file
label. Arguably the per-pane location is much closer to the original
intention of the --label argument, but either way we need to have some
use of the provided labels in the relatively normal situation where a
single pane comparison has been opened from e.g., a VC difftool launch.