Move items from horizontal-icon section to regular menu items
The horizontal icon section doesn't really establish a particularly
coherent grouping of functions, and this style is not widely used enough
for it to be familiar to all users.

We've just moved items to relevant submenus, except for Refresh which
we're giving additional prominence because it's still needed somewhat
regularly for certain use cases.