filediff: Ignore cursor changes for non-focused panes (#646)
Certain actions (such as pushing a chunk to a pane) manually trigger our
cursor change logic, which means that we can get cursor change
notifications for panes that aren't focused. This was causing confusion
because actions such as next/previous chunk would check the cursor
line number and get the line number from a non-focused pane.

The fix here is fairly simple; we just don't run our cursor-changed
logic if the cursor change happens in a pane that isn't focused.