Change warning silencing logic
This switches our logic about whether to silence warnings from being
based on version number to being based on whether we're running from
Git. Either of these is probably going to be fine for us, but using
.git has the benefit that it will work even if we change away from the
odd/even:unstable/stable versioning scheme, and also means that we don't
need to have `meld.conf` available by the time we do this check.