accelerators: Fix go-to-line accelerator conflict (#729)
We were using Ctrl-I for both go-to-line and previous-conflict. Under
normal circumstances, I would prioritise the older accelerator usage,
but since go-to-line was winning out and has been for the entire 3.20.x
series without anyone complaining, it's hard to imagine that it's been
very missed.

Given that Ctrl-I is used as the go-to-line shortcut in other GNOME
text editor applications, we'll prioritise consistency here and change
previous-conflict instead.