Wait for index update operation in git VC interactions (#734)
In general, the index update operation is fast enough (and the index
up-to-date enough) that this isn't a problem. However, in certain
situations git takes long enough to update the index that the subsequent
diff-index operations will get stale data. The fix here is that we wait
for the update-index operation to complete before continuing on to the
diff-index, etc. checks.

That we didn't do this previously was almost certainly a simple
oversight; other calls effectively wait because they rely on stdout
results.