ui.msgarea: Work around layout sizing issue (#713)
This is obviously nonsensical, but there appears to be a bug in GTK's
sizing where having *two* labels with word+char wrapping in the same box
triggers some kind of incorrect height calculation that flows through to
the minimum sizing of the container somehow, with the symptom being that
if you get an infobar notification that needs to wrap the second box,
your window now isn't resizable and is (usually) taller than the screen.

I couldn't reproduce this until a recent distribution upgrade, and there
are some indications that there might be something going on with a newer
Cairo, or possibly interactions between it and font sizing calculations?

Regardless, this is a weird bug that I can't reproduce if we only make
one of the two labels WRAP_WORD_CHAR. Here, we're making the second
label word-only wrap because the first label is the one that usually
contains e.g., filenames, which were the original reason for allowing
character wrapping (which is not in general necessary).

Writing a minimal reproducer for this is not so far going well.