Don't save and restore fullscreen state (#716)
While trying to keep window state is generally a good idea, in the case
of fullscreen state it's not obviously true. Looking at several current
GNOME applications, most of them don't support fullscreen at all, and of
the ones that do, I can't find any that save and restore the fullscreen
state across application restart.

This change is triggered by the problems with our fullscreen UI
described in #716, but it's really mostly independent of actually fixing
those issues.