Avoid popover focus-out triggers on view refreshes (#774)
The problem we're addressing here is that focus-grabbing menus such as
our file/text/state filter popovers will close if another widget grabs
focus. In some cases this is fine, but in the case of our filter menus
it actually makes sense to allow the user to make multiple selections at
once.

We were grabbing focus mostly to re-establish selection state, but also
to ensure that the user had a correctly focused pane when they opened
a new comparison. In testing, the first of these is still a problem in
e.g., file comparisons with changing action sensitivity. The second of
these has more-or-less gone away with changing GTK versions and our
layout updates. In particular, the full refreshes required by folder and
version control comparisons mean that we reestablish the state we need
without much additional work.