Fix incorrect file comparison copy-chunk sensitivity
In a refactor these brackets were incorrect removed. Copy actions are
not supposed to be enabled for insert or delete chunks, and this makes
that true. For chunk copies that at least targeted the correct pane,
we would previously (incorrectly) do basically the same thing as a move.
However, the incorrect logic also enabled these actions on non-existent
panes, which resulted in tracebacks.