Update cut/copy/paste sensitivity on pane change (#780)
We previously only updated this on startup (when we have no focused
pane) and when the selection changed. This was most obviously wrong in
that it required selecting something before paste was available, however
it was also bad for cut + copy because the user could select something
in the left pane, then right click in the right pane, and cut + copy
would still be sensitive.