Add F8 shortcut for filter menu buttons (#775)
This fills in a missing piece of the new filter UI, making it actually
keyboard accessible.

The F8 shortcut isn't necessarily final, and for now the shortcut is
shared among all three filter buttons. This is probably how it will
remain, since it's probably the best option for common usage, although
it does pose an issue for adding the text filter UI to folder
comparisons.