Reorganise sections in the shortcut window (#793)
The General section was very long, and contained several items that were
definitely not actually general in nature. This commit splits general
in to three sections: General, Window and Tab, and Common Actions. This
also helps address the issue raised in #793 that we didn't have the menu
shortcut listed. It turns out we did, but it was buried in the middle of
a bunch of unrelated options; it should now be more obvious.