Try to disable our spinner on X11 forwarding (#507)
This is a bit weak, since detecting X11 forwarding doesn't appear to be
generally viable, but also if we're wrong then the worst case is that
the spinner doesn't spin, so it's probably not a massive issue.

The reason we want to do this is that paint priority and timing
interferes with us doing most of our calculations on idle, which means
that we spend all our time spinning and don't actually do any work.