Add check for white-space-varying file names (#808)
This check addresses cases where two folders have files that would have
the same name but for leading or trailing white space, which isn't the
kind of thing that is really user visible otherwise. While a user
*might* notice leading white space, there's basically no chance that
they'll notice trailing white space.

Importantly, we don't try to alert merely on the presence of leading or
trailing white space; we deliberately only detect if files would be
named the same if it weren't for the leading/trailing space.