Restyle path label to show that it's not editable (#260)
For reasons that are unclear to me, the editability of the entry isn't
causing a useful restyle. The easiest fix here appears to be to make it
look like not an entry, since we're using the entry widget here for its
secondary characteristics (i.e., being single-line scrollable and
supporting keynav) rather than to support any kind of actual editing.

The best way I found to do this was to give it the flat styling which
makes it very label-like, and then to also make sure it doesn't
immediately get keyboard focus when the popover opens (by allowing its
form label to get focus instead).