Make file compare-to-itself check more accurate (#856)
The original bug here is that admin:/// URIs don't have a path, and so
were incorrectly triggering the duplicate comparison check. However,
doing a URI comparison is also wrong because we still want to warn if
e.g., the user is comparing a file:/// URI to the same file with a
relative path.

The fix here is to do the check based on Gio's file ID. This appears to
handle the above situations correctly, as well as handling symlinks that
resolve to the same file, etc.