Mostly port MeldWindow to GTK 4
There are several minor bits remaining (including some changes to
NewDiffTab for it to display at all) but this is the vast majority of
the changes. The main things remaining are the menu handling for the
various popovers, and the fact that we've had to comment out the entire
recent files widget.