Fix incorrect parent state in path label when opening new files (#881)
If the user opened a file in an existing comparison that would change
the shared-parent calculation, PathLabel didn't handle the change and
refused to set its path because the new path didn't match the calculated
parent. Ideally PathLabel would figure out the parent situation itself,
but that requires PathLabel widgets knowing about each other, and that's
a bigger change than this.