Remove Ctrl+Page Up/Down bindings for horizontal scroll (#879)
We want to use these for tab switching instead of horizontal scroll,
since that's a much more useful activity in Meld in general. Also, we've
had that as the documented and used shortcut for ages now, so I'm not
willing to change it.

The effect here is that instead of these shortcuts working sometimes
(i.e., in file comparisons, or in folder/VC comparisons when you didn't
have the treeview focused) they should now work consistently.