Widen exception handling when checking requirements
This is prompted by a current issue in the Windows build where importing
GdkPixbuf gobject introspection raises an assertion error, which we
don't catch and then silently fail to launch. While ideally we wouldn't
show a traceback here, it's probably more important to do that than to
nicely fail, since in any normal install this shouldn't be a problem.