Replace CSS key binding set with shortcut-only action
The goal here is to get an error bell when we try to go past the
first/last change in the file using the keyboard, as a piece of useful
user feedback. The previous approach used a signal not attached to the
action, but the keybinding approach doesn't work in CSS any more. We
could migrate to actually constructing binding sets in code or in the
UI files, but the approach here - just having a duplicate action that
doesn't get affected by sensitivity setting - seems a lot easier.