Move Filediff right-click controller to UI template
This is way trickier than it should be because GTK 4 has removed the
the ways that we'd otherwise be popping up a menu, and there's no way
to suppress the default GtkTextView context menu without capturing the
event in the propagation phase. This is a moderately gross hack, but
this is the only option I could find.