Remove the Filediff scrollwindow size allocation handler
This was previously needed for (from memory) the situation where the
allocation of one pane's window changed, typically because of the
presentation of a GtkInfoBar that moved the text view down. Because the
linkmap widget didn't respond to this, the link diff drawing would no
longer line up with the text view diff chunks. In testing, this no
longer seems to be a problem (not that surprising given how many
iterations these widgets have been through); if it resurfaces, this is
fairly easy to add back.